requirements.txt
pandas==1.4.4
Pillow==9.2.0
scikit_learn==1.1.2
streamlit==1.12.2
